Refactoring of ManageOtc and ManageDrug pages
